Remove dead code
This snippet of code occurs twice, so the second appearance of this condition is impossible to be satisfied (X1/X2/Y1/Y2 are not modified in the anonymous function). So this code is dead. It does nothing and never will. Which is lucky, because the anonymous function is supposed to return a number, not a function (`Linear`).